refactor(db, parser, ci): post-review improvements
- Replace flogging with logging.New; ban flogging via depguard
- Fix DSN via dbconn.DataSourceName; fix rollback to use context.Background()
- Remove ProcessedBlock.Txns and TxRecord.BlockNum; derive at write time
- Move buildBatchParams before beginTx; add nil guards for BlockInfo and Data
- Pre-allocate all seven batchParams slices
- Change TxRecord.ValidationCode from int32 to protoblocktx.Status (typed)
- Change validation_code column from BIGINT to SMALLINT
- Log warning when endorsement identity parsing fails (was silently swallowed)
- Add util.Ptr[T] generic helper; remove scattered ptr helpers
- Rename test_exports.go to test_exports_test.go
- Add TestMain with DB_DEPLOYMENT=local and DB_TYPE=postgres
- CI: add lint and test jobs; pin all action SHAs; add permissions block;
  set timeout-minutes; pin golangci-lint to v2.10.1; verify sqlc SHA256
- Makefile: add start-db/stop-db targets

+++ pkg/db/sqlc/models.go
@@ -25,7 +25,7 @@ type Transaction struct {
 	BlockNum       int64  `json:"block_num"`
 	TxNum          int64  `json:"tx_num"`
 	TxID           []byte `json:"tx_id"`
-	ValidationCode int64  `json:"validation_code"`
+	ValidationCode int16  `json:"validation_code"`
 }
 
 type TxBlindWrite struct {
